# api_docs/api-docs-input.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest
      summary: Returns a specific user
      operationId: profile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User data returned correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
    put:
      tags:
      - user-rest
      summary: Edits a user profile
      operationId: editProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "403":
          description: Operation not allowed for the current user
          content:
            application/json: {}
        "400":
          description: Exception originated from JSON data processing or mapping
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "200":
          description: User profile edited correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
    delete:
      tags:
      - user-rest
      summary: Deletes a user profile
      operationId: deleteProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Operation successful
        "403":
          description: Operation only allowed to admins or the user itself
          content:
            application/json: {}
        "404":
          description: Profile not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE'
...
  /api/reviews/{reviewId}/users/{userId}:
    put:
      tags:
      - review-rest
      summary: Edit a review
      description: Edit a review by review ID and user ID.
      operationId: editReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "404":
          description: User or review not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "200":
          description: Review edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
...
  /api/reviews/hotels/{hotelId}:
    post:
      tags:
      - review-rest
      summary: Post a hotel review
      description: Post a review for a specific hotel by hotel ID.
      operationId: postReview
      parameters:
      - name: hotelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "400":
          description: Invalid input
          content:
            application/json: {}
        "201":
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review'
...
  /api/users/reviews/{id}:
    get:
      tags:
      - user-rest
      summary: Returns a specific user from a specific review
      operationId: reviewUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "200":
          description: User data returned correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserE_UserDetails'
        "403":
          description: Operation only allowed for admins or the user that made the
            review
          content:
            application/json: {}
...
  /api/reviews:
    get:
      tags:
      - review-rest
      summary: Load all reviews
      description: Load all reviews with pagination.
      operationId: loadAllReviews
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_HotelDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReview_HotelDetails'
  /api/reviews/{id}:
    get:
      tags:
      - review-rest
      summary: Get a review
      description: Get a specific review by its ID.
      operationId: getReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Review retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
        "404":
          description: Review  not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review_ReviewDetails'
  /api/reviews/users/{id}:
    get:
      tags:
      - review-rest
      summary: Get user reviews
      description: Get all reviews of a specific user by user ID.
      operationId: userReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReviewDetails'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReview_ReviewDetails'
  /api/reviews/size/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get the number of hotel reviews
      description: Get the number of reviews of a specific hotel by hotel ID.
      operationId: hotelReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "200":
          description: Number of reviews retrieved successfully
          content:
            application/json:
              schema:
                type: integer
                format: int32
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: integer
                format: int32
  /api/reviews/percentage/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get the percentage of hotel reviews
      description: Get the percentage of reviews of a specific hotel by hotel ID.
      operationId: hotelReviews_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Percentage of reviews retrieved successfully
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: integer
                  format: int32
  /api/reviews/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get hotel reviews
      description: Get all reviews of a specific hotel by hotel ID.
      operationId: hotelReviews_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReviewDetails'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReview_ReviewDetails'
...
  /api/hotels/reviews/{id}:
    get:
      tags:
      - hotel-rest
      summary: Get reviews from a hotel
      description: Returns a list of reviews for a specific hotel.
      operationId: getReviewsFromHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: "Forbidden, the user does not have permission to view these\
            \ reviews"
          content:
            application/json: {}
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Hotel_HotelDetails'
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review_HotelDetails'
...
  /api/reviews/{reviewId}:
    delete:
      tags:
      - review-rest
      summary: Delete a review
      description: Delete a review by review ID.
      operationId: deleteReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "204":
          description: Review deleted successfully
          content:
            application/json: {}
components:
  schemas:
    Hotel_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        rooms:
          $ref: '#/components/schemas/Room_UserDetails'
        idReservation:
          type: integer
          format: int64
    Review_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        idReview:
          type: integer
          format: int64
...
    UserE_UserDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_UserDetails'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review_UserDetails'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/Hotel_UserDetails'
        imageBase64:
          type: string
...
    Hotel_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Review_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel_ReviewDetails'
        user:
          $ref: '#/components/schemas/UserE_ReviewDetails'
        idReview:
          type: integer
          format: int64
...
    Hotel_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        manager:
          $ref: '#/components/schemas/UserE_HotelDetails'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room_HotelDetails'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_HotelDetails'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review_HotelDetails'
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
...
    Review_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        idReview:
          type: integer
          format: int64
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        hotel:
          $ref: '#/components/schemas/Hotel'
        room:
          $ref: '#/components/schemas/Room'
        user:
          $ref: '#/components/schemas/UserE'
        rooms:
          $ref: '#/components/schemas/Room'
        idReservation:
          type: integer
          format: int64
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel'
        user:
          $ref: '#/components/schemas/UserE'
        idReview:
          type: integer
          format: int64
...
    UserE:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/Hotel'
        imageBase64:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageResponse_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponseReview_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review_HotelDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponse_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseReview_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review_ReviewDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...